smoothly

+++ src/store/modules/report.js
@@ -24,17 +24,29 @@ const actions = {
   getAllReportData ({ commit, state, dispatch }) {
     const filterType = state.filterType
     dispatch('loading/openLoading', null, { root: true })
-    AppService.getReports({ filterType }).then(res => {
+    AppService.getReports({ filterType })
+      .then(res => {
       commit('setReportData', res.data)
-      dispatch('loading/closeLoading', null, { root: true })
+    }).catch(() => {
+      Vue.$toast.error('Get data failed')
     })
+      .finally(() => {
+        dispatch('loading/closeLoading', null, { root: true })
+      })
   },
   updateReportFields (context, updateData) {
     context.dispatch('loading/openLoading', null, { root: true })
-    AppService.updateReportFields(updateData).then(res => {
+    AppService.updateReportFields(updateData)
+      .then(res => {
       context.commit('updateReport', res.data)
-      context.dispatch('loading/closeLoading', null, { root: true })
+      Vue.$toast.success('Update Success')
     })
+      .catch(() => {
+        Vue.$toast.error('Update Failed')
+      })
+      .finally(() => {
+        context.dispatch('loading/closeLoading', null, { root: true })
+      })
   },
   resetReportData (context) {
     context.dispatch('loading/openLoading', null, { root: true })
@@ -51,17 +63,39 @@ const actions = {
     const filterType = state.filterType
     const messageTelegram = state.messageTelegram
     dispatch('loading/openLoading', null, { root: true })
-    AppService.sendMessageToTelegram({ filterType }, { message: messageTelegram }).then(res => {
-      dispatch('loading/closeLoading', null, { root: true })
+    AppService.sendMessageToTelegram({ filterType }, { message: messageTelegram })
+      .then(res => {
       dispatch('dialog/closeDialogSendMessageTelegram', null, { root: true })
+        Vue.$toast.success('Send message to telegram success')
+    }).catch(() => {
+      Vue.$toast.error('Send message to telegram failed')
+    }).finally(() => {
+      dispatch('loading/closeLoading', null, { root: true })
     })
   },
   deleteReportByAccountId (context, { accountId }) {
-    context.dispatch('loading/openLoading', null, { root: true })
-    AppService.deleteReportByAccountId(accountId).then(res => {
-      context.commit('removeReport', res.data)
-      context.dispatch('loading/closeLoading', null, { root: true })
-    })
+    Vue.$confirm(
+        {
+          message: 'Are you sure delete item?',
+          button: {
+            no: 'Cancel',
+            yes: 'OK',
+          },
+          callback: confirm => {
+            if (confirm) {
+              context.dispatch('loading/openLoading', null, { root: true })
+              AppService.deleteReportByAccountId(accountId).then(res => {
+                context.commit('removeReport', res.data)
+                Vue.$toast.success('Delete item success')
+              }).catch((e) => {
+                Vue.$toast.error('Delete item failed')
+              }).finally(() => {
+                context.dispatch('loading/closeLoading', null, { root: true })
+              })
+            }
+          },
+        },
+      )
   },
 }
 const getters = {
